Completed modals + file sort
