+++ apps/iot-app/src/Icons/AlertCanceledIcon.tsx
@@ -1,0 +1,18 @@
+const AlertCanceledIcon = () => (
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    width="64"
+    height="64"
+    fill="none"
+    stroke="currentColor"
+    strokeWidth="2"
+    strokeLinecap="round"
+    strokeLinejoin="round"
+    className="text-blue-500"
+    viewBox="0 0 24 24"
+  >
+    <line x1="18" y1="6" x2="6" y2="18" />
+    <line x1="6" y1="6" x2="18" y2="18" />
+  </svg>
+);
+export default AlertCanceledIcon;
