+++ apps/iot-app/src/Icons/LostConnectionIcon.tsx
@@ -1,0 +1,20 @@
+const LostConnectionIcon = () => (
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    width="64"
+    height="64"
+    fill="none"
+    stroke="currentColor"
+    strokeWidth="2"
+    strokeLinecap="round"
+    strokeLinejoin="round"
+    className="text-blue-600"
+    viewBox="0 0 24 24"
+  >
+    <path d="M16 8a6 6 0 0 0-8 0" />
+    <path d="M12 20v.01" />
+    <path d="M12 16v.01" />
+    <path d="M12 12v.01" />
+  </svg>
+);
+export default LostConnectionIcon;
