+++ apps/iot-app/src/Icons/LowBatteryIcon.tsx
@@ -1,0 +1,18 @@
+const LowBatteryIcon = () => (
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    width="64"
+    height="64"
+    fill="none"
+    stroke="currentColor"
+    strokeWidth="2"
+    strokeLinecap="round"
+    strokeLinejoin="round"
+    className="text-red-400"
+    viewBox="0 0 24 24"
+  >
+    <rect x="2" y="7" width="18" height="10" rx="2" ry="2" />
+    <line x1="22" y1="11" x2="22" y2="13" />
+  </svg>
+);
+export default LowBatteryIcon;
