+++ apps/iot-app/src/pages/MainPage/MainPage.tsx
@@ -1,5 +1,7 @@
 import React, { useState } from "react";
 
+import ArrowLeftIcon from "../../Icons/ArrowLeftIcon";
+import ArrowRightIcon from "../../Icons/ArrowRightIcon";
 import Alert from "../../alerts/Alert";
 import { useTheme } from "../../functions/ThemeContext";
 
@@ -27,39 +29,7 @@ export default function MainPage() {
           onClick={() => setIsSidebarOpen(prev => !prev)}
           className="absolute left-2 top-2 z-10 bg-gray-600 text-white px-3 py-1 rounded hover:bg-gray-800 hover:shadow-xs hover:shadow-gray-600/50 transition"
         >
-          {isSidebarOpen ? (
-            <svg
-              xmlns="http://www.w3.org/2000/svg"
-              width="24"
-              height="24"
-              viewBox="0 0 24 24"
-              fill="none"
-              stroke="currentColor"
-              stroke-width="2"
-              stroke-linecap="round"
-              stroke-linejoin="round"
-              className="feather feather-arrow-left"
-            >
-              <line x1="19" y1="12" x2="5" y2="12"></line>
-              <polyline points="12 19 5 12 12 5"></polyline>
-            </svg>
-          ) : (
-            <svg
-              xmlns="http://www.w3.org/2000/svg"
-              width="24"
-              height="24"
-              viewBox="0 0 24 24"
-              fill="none"
-              stroke="currentColor"
-              stroke-width="2"
-              stroke-linecap="round"
-              stroke-linejoin="round"
-              className="feather feather-arrow-right"
-            >
-              <line x1="5" y1="12" x2="19" y2="12"></line>
-              <polyline points="12 5 19 12 12 19"></polyline>
-            </svg>
-          )}
+          {isSidebarOpen ? <ArrowLeftIcon /> : <ArrowRightIcon />}
         </button>
 
         <button
