+++ apps/iot-app/src/pages/MainPage/SearchBar.tsx
@@ -1,5 +1,6 @@
 import React from "react";
 
+import FilterIcon from "../../Icons/FilterIcon";
 import { useTheme } from "../../functions/ThemeContext";
 
 const SearchBar: React.FC = () => {
@@ -20,20 +21,7 @@ const SearchBar: React.FC = () => {
         <button
           className={`ml-2 p-2 rounded ${bgColor} ${textColor} border ${borderColor} hover:bg-opacity-80 transition`}
         >
-          <svg
-            xmlns="http://www.w3.org/2000/svg"
-            width="20"
-            height="20"
-            viewBox="0 0 24 24"
-            fill="none"
-            stroke="currentColor"
-            strokeWidth="2"
-            strokeLinecap="round"
-            strokeLinejoin="round"
-            className="feather feather-filter"
-          >
-            <polygon points="22 3 2 3 10 12.46 10 19 14 21 14 12.46 22 3" />
-          </svg>
+          <FilterIcon />
         </button>
       </div>
     </div>
